refactor(nostr): extract NostrContext module-scope helpers into sibling modules
Pure structural split — move module-scope helpers out of NostrContext.tsx
(4279 → 3565 lines) into cohesive sibling modules, zero behaviour change.
The React provider + all hooks stay in NostrContext.tsx; NostrContextType
stays too (tightly coupled to the component).

New modules under src/contexts/:
- nostrAuthKeys.ts      — NSEC_KEY / PUBKEY_KEY / SIGNER_TYPE_KEY
- nostrContextTypes.ts  — SignedEvent / ConversationMessage / RefreshDmInboxOptions
- nostrEventBus.ts      — group + DM pub/sub (notify/subscribe + listener sets)
- nostrGroupRouting.ts  — tryRouteGroupRumor + GroupRouteResult
- nostrDecryptPacing.ts — yieldToEventLoop, createYieldScheduler, yield cadences
- nostrSecretKeyCache.ts— nip04 plaintext + memoised secret-key caches
- nostrDmCache.ts       — NIP-17 wrap cache + DM inbox/conversation cache
- nostrCacheKeys.ts     — profile/contacts/relay cache keys + readCachedWithTtl

Externally-imported symbols (OWN_PROFILE_CACHE_KEY_BASE, notifyGroupMessage,
subscribeGroupMessages, subscribeDmMessages, the shared types) are re-exported
from NostrContext.tsx so existing importers stay untouched.

Gates: tsc clean, jest 700/700 (unchanged), eslint 0 errors (111 warnings,
unchanged from main), prettier clean, file-size check passes.
