+++ src/contexts/nostrAuthKeys.ts
@@ -1,0 +1,3 @@
+export const NSEC_KEY = 'nostr_nsec';
+export const PUBKEY_KEY = 'nostr_pubkey';
+export const SIGNER_TYPE_KEY = 'nostr_signer_type';
